refactor form config for maintainability
