+++ app/types/signupForm.ts
@@ -1,0 +1,1 @@
+export type ValidationFunction = (value: string) => string | null;
